# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IhaleTakip API
  version: 1.0.0
  description: Kamu ihalesi takip uygulaması backend servisi
paths:
...
  /api/v1/ai/tasks/{task_id}/:
    get:
      operationId: api_v1_ai_tasks_retrieve
      description: '`POST /ai/analyze` yanıtındaki `task_id` ile görevin durumunu
        döner.


        | `data.status` | HTTP | Anlamı |

        |---|---|---|

        | `pending` | 200 | Kuyrukta bekliyor |

        | `processing` | 200 | Worker işliyor |

        | `completed` | 200 | `data.analysis` ve `data.usage` dolu |

        | `failed` | 422 / 500 | `message` hatayı açıklar |


        İstemci `completed` veya `failed` görene kadar birkaç saniye aralıkla yoklar.'
      summary: Analiz durumunu sorgula
      parameters:
      - in: path
        name: task_id
        schema:
          type: string
        description: '`POST /ai/analyze` yanıtındaki Celery görev kimliği (UUID).'
        required: true
        examples:
          GörevKimliği:
            value: 3f8c2b1a-7e4d-4a91-9c2f-8b6d5e1a0c33
            summary: Görev kimliği
      tags:
      - ai
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResult'
          description: ''
...
  /api/v1/assistant/conversations/{id}/:
    get:
      operationId: api_v1_assistant_conversation_messages
      description: Seçili oturumun mesajlarını **en yeniden eskiye** döner (inverted
        liste).
      summary: Sohbet oturumunun mesajlarını listele
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      - in: query
        name: page
        schema:
          type: integer
        description: Sayfa numarası.
      - in: query
        name: page_size
        schema:
          type: integer
        description: Sayfa boyutu.
      tags:
      - assistant
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationMessagePage'
          description: ''
    delete:
      operationId: api_v1_assistant_conversations_destroy
      description: Oturumu ve içindeki tüm mesajları siler.
      summary: Sohbet oturumunu sil
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - assistant
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDeleted'
          description: ''
...
components:
  schemas:
    AnalysisResult:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/AnalysisResultStatusEnum'
        cached:
          type: boolean
        analysis: {}
        usage: {}
      required:
      - status
    AnalysisResultStatusEnum:
      enum:
      - pending
      - processing
      - completed
      - failed
      type: string
      description: '* `pending` - pending

        * `processing` - processing

        * `completed` - completed

        * `failed` - failed'
...
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        conversation:
          type: integer
          readOnly: true
          nullable: true
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          readOnly: true
        content:
          type: string
          readOnly: true
        payload:
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - content
      - conversation
      - created_at
      - id
      - payload
      - role
...
    ConversationDeleted:
      type: object
      properties:
        deleted:
          type: integer
      required:
      - deleted
    ConversationMessagePage:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
        totalCount:
          type: integer
        page:
          type: integer
      required:
      - list
      - page
      - totalCount
...
    RoleEnum:
      enum:
      - user
      - assistant
      type: string
      description: '* `user` - Kullanıcı

        * `assistant` - Asistan'
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT